requirements.txt
altair
pandas
streamlit
mysql-connector-python
openpyxl
